+++ Controls/_popup/Opener/BaseOpener.ts
@@ -258,23 +258,50 @@ class BaseOpener<TBaseOpenerOptions extends IBaseOpenerOptions = {}>
 
         if (!BaseOpener.isNewEnvironment()) {
             BaseOpener.getManager().then(() => {
-                BaseOpener.getZIndexUtil().addCallback((getZIndex) => {
-                    if (popupOpener) {
-                        // при открытии через стат. метод открыватора в верстке нет, нужно взять то что передали в опции
-                        // Если topPopup, то zIndex менеджер высчитает сам
+                if (popupOpener) {
+                    // при открытии через стат. метод открыватора в верстке нет, нужно взять то что передали в опции
+                    // Если topPopup, то zIndex менеджер высчитает сам
+
+                    if (!cfg.zIndex) {
                         if (!cfg.topPopup) {
-                            cfg.zIndex = cfg.zIndex || getZIndex(popupOpener);
+                            // На старой странице могут открывать на одном уровне 2 стековых окна.
+                            // Последнее открытое окно должно быть выше предыдущего, для этого должно знать его zIndex.
+                            // Данные хранятся в WM
+                            const oldWindowManager = requirejs('Core/WindowManager');
+                            const compatibleManagerWrapperName = 'Controls/Popup/Compatible/ManagerWrapper/Controller';
+                            let managerWrapperMaxZIndex = 0;
+                            // На старой странице может быть бутерброд из старых и новых окон. zIndex вдомных окон берем
+                            // из менеджера совместимости. Ищем наибольший zIndex среди всех окон
+                            if (requirejs.defined(compatibleManagerWrapperName)) {
+                               managerWrapperMaxZIndex = requirejs(compatibleManagerWrapperName).default.getMaxZIndex();
+                            }
+                            const zIndexStep = 9;
+                            const item = ManagerController.find(cfg.id);
+                            // zindex окон, особенно на старой странице, никогда не обновлялся внутренними механизмами
+                            // Если окно уже открыто, zindex не меняем
+                            if (item) {
+                                cfg.zIndex = item.popupOptions.zIndex;
+                            } else if (oldWindowManager) {
+                                // Убираем нотификационные окна из выборки старого менеджера
+                                const baseOldZIndex = 1000;
+                                const oldMaxZWindow = oldWindowManager.getMaxZWindow((control) => {
+                                    return control._options.isCompoundNotification !== true;
+                                });
+                                const oldMaxZIndex = oldMaxZWindow?.getZIndex() || baseOldZIndex;
+                                const maxZIndex = Math.max(oldMaxZIndex, managerWrapperMaxZIndex);
+                                cfg.zIndex = maxZIndex + zIndexStep;
+                            }
                         }
                     }
-                    if (!BaseOpener.isVDOMTemplate(rootTpl)) {
-                        requirejs(['Controls/compatiblePopup'], (compatiblePopup) => {
-                            compatiblePopup.BaseOpener._prepareConfigForOldTemplate(cfg, rootTpl);
-                            BaseOpener._openPopup(cfg, controller, def);
-                        });
-                    } else {
+                }
+                if (!BaseOpener.isVDOMTemplate(rootTpl)) {
+                    requirejs(['Controls/compatiblePopup'], (compatiblePopup) => {
+                        compatiblePopup.BaseOpener._prepareConfigForOldTemplate(cfg, rootTpl);
                         BaseOpener._openPopup(cfg, controller, def);
-                    }
-                });
+                    });
+                } else {
+                    BaseOpener._openPopup(cfg, controller, def);
+                }
             });
         } else if (BaseOpener.isVDOMTemplate(rootTpl) &&
             !(cfg.templateOptions && cfg.templateOptions._initCompoundArea)) {
