+++ Controls/_popup/Opener/Stack.ts
@@ -45,36 +45,8 @@ const getStackConfig = (stackOptions: IStackOpenerOptions = {}, popupId?: string
     // For more information, see  {@link Controls/interface/ICanBeDefaultOpener}
     config.isDefaultOpener = config.isDefaultOpener !== undefined ? config.isDefaultOpener : true;
     config._type = 'stack'; // TODO: Compatible for compoundArea
+    config.id = config.id || popupId;
 
-    // TODO: Compatible
-    // На старой странице могут открывать на одном уровне 2 стековых окна.
-    // Последнее открытое окно должно быть выше предыдущего, для этого должно знать его zIndex. Данные хранятся в WM
-    if (!isNewEnvironment() && !config.zIndex) {
-        const oldWindowManager = requirejs('Core/WindowManager');
-        const compatibleManagerWrapperName = 'Controls/Popup/Compatible/ManagerWrapper/Controller';
-        let managerWrapperMaxZIndex = 0;
-        // На старой странице может быть бутерброд из старых и новых окон. zIndex вдомных окон берем
-        // из менеджера совместимости. Ищем наибольший zIndex среди всех окон
-        if (requirejs.defined(compatibleManagerWrapperName)) {
-            managerWrapperMaxZIndex = requirejs(compatibleManagerWrapperName).default.getMaxZIndex();
-        }
-        const zIndexStep = 9;
-        const item = ManagerController.find(config.id || popupId);
-        // zindex окон, особенно на старой странице, никогда не обновлялся внутренними механизмами
-        // Если окно уже открыто, zindex не меняем
-        if (item) {
-            config.zIndex = item.popupOptions.zIndex;
-        } else if (oldWindowManager) {
-            // Убираем нотификационные окна из выборки старого менеджера
-            const baseOldZIndex = 1000;
-            const oldMaxZWindow = oldWindowManager.getMaxZWindow((control) => {
-                return control._options.isCompoundNotification !== true;
-            });
-            const oldMaxZIndex = oldMaxZWindow?.getZIndex() || baseOldZIndex;
-            const maxZIndex = Math.max(oldMaxZIndex, managerWrapperMaxZIndex);
-            config.zIndex = maxZIndex + zIndexStep;
-        }
-    }
     return config;
 };
 const POPUP_CONTROLLER = 'Controls/popupTemplate:StackController';
