+++ Controls/_popupTemplate/Dialog/Opener/DialogController.ts
@@ -1,10 +1,8 @@
 import {default as BaseController, IDragOffset} from 'Controls/_popupTemplate/BaseController';
 import {IPopupItem, IPopupOptions, IPopupSizes, IPopupPosition} from 'Controls/popup';
 import {detection} from 'Env/Env';
-import * as isNewEnvironment from 'Core/helpers/isNewEnvironment';
 import * as Deferred from 'Core/Deferred';
 import DialogStrategy = require('Controls/_popupTemplate/Dialog/Opener/DialogStrategy');
-import oldWindowManager from 'Controls/_popupTemplate/_oldWindowManager';
 
 interface IDialogItem extends IPopupItem {
     popupOptions: IDialogOptions;
@@ -40,9 +38,6 @@ class DialogController extends BaseController {
 
     elementCreated(item: IDialogItem, container: HTMLDivElement): boolean {
         this._prepareConfigWithSizes(item, container);
-        if (!isNewEnvironment()) {
-            oldWindowManager.addZIndex(item.currentZIndex);
-        }
         return true;
     }
 
@@ -86,9 +81,6 @@ class DialogController extends BaseController {
     }
 
     elementDestroyed(item: IPopupItem): Promise<null> {
-        if (!isNewEnvironment()) {
-            oldWindowManager.removeZIndex(item.currentZIndex);
-        }
         return (new Deferred()).callback();
     }
